fix: address PR review feedback for bundle command
- Extract readStdin to shared utility in src/utils/options.ts
- Check for commas before existsSync to reduce key/file ambiguity
- Document file-detection behavior in --keys description
