+++ src/lib/iam/policies/utils.ts
@@ -1,12 +1,6 @@
 import type { PolicyDocument } from '@tigrisdata/iam';
 
-export async function readStdin(): Promise<string> {
-  const chunks: Buffer[] = [];
-  for await (const chunk of process.stdin) {
-    chunks.push(chunk);
-  }
-  return Buffer.concat(chunks).toString('utf-8');
-}
+export { readStdin } from '@utils/options.js';
 
 export function parseDocument(jsonString: string): PolicyDocument {
   const raw = JSON.parse(jsonString);
